api/requirements.txt
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
django-debug-toolbar==4.2.0
channels==4.0.0
psycopg2==2.9.9
redis==5.0.1
daphne==4.0.0
jsonschema>4.0.0